feat: add react-copy-to-clipboard package and use it in CoverLetter.tsx
This commit adds the react-copy-to-clipboard package as a dependency
and uses it to copy the cover letter content to the clipboard when
the user clicks the button. It also adds some environment variables
to store the user's name, mobile and email, and uses them to generate
the cover letter content. It also updates the .gitignore file to
ignore the .env file. (main)